backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
Django==4.2.18
djangorestframework==3.15.2
docstring_parser==0.16
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
instructor==1.7.2
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.14
langchain-core==0.3.29
langchain-openai==0.3.0
langchain-text-splitters==0.3.5
langsmith==0.2.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
numpy==1.26.4
openai==1.59.7
orjson==3.10.14
packaging==24.2
propcache==0.2.1
psycopg2-binary==2.9.9
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlparse==0.5.3
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
